// src/types/keyCode.ts
export const enum KeyCode {
  Up = 'ArrowUp',
  Down = 'ArrowDown',
  Left = 'ArrowLeft',
  Right = 'ArrowRight',
  PgUp = 'PageUp',
  PgDown = 'PageDown',
  Home = 'Home',
  End = 'End',
  Enter = 'Enter',
  Alt = 'Alt',
  Ctrl = 'Control',
  Shift = 'Shift',
  Meta = 'Meta',
  Space = ' ',
  Tab = 'Tab',
  Esc = 'Escape',
  Del = 'Delete',
  Backspace = 'Backspace',
}